api-gateway/requirements.txt
ccxt
requests
Flask
kafka-python
flask-cors

